dynamically accessing blog Detail page through slug
