+++ components/blog/blogSubSection2.js
@@ -12,7 +12,7 @@ const BlogSubSection2 = ({ posts }) => {
         >
           <div className="card-blog-grid card-blog-grid-3 hover-up">
             <div className="card-image">
-              <Link href="/blog-detail">
+              <Link href={post.slug}>
                 <FeaturedImage post={post} styleClasses="" priority={false} />
                 {/* <Image
                 layout="responsive"
@@ -28,7 +28,7 @@ const BlogSubSection2 = ({ posts }) => {
               </Link>
             </div>
             <div className="card-info">
-              <Link href="/blog-detail">
+              <Link href={post.slug}>
                 <h4 className="color-brand-1">
                   {post.title}
                   {/* How to save money - 8 simple ways to start saving money every
