+++ components/cards/BlogCard.js
@@ -3,7 +3,7 @@ import Image from "next/image";
 import FeaturedImage from "../elements/FeaturedImage";
 import Date from "../elements/Date";
 
-const BlogCard = ({ title, date, readTimeTag, excerpt, tag, post }) => {
+const BlogCard = ({ title, date, readTimeTag, excerpt, tag, slug, post }) => {
   return (
     <>
       <div className={`col-lg-4 col-md-6 mb-30 item-article ${tag}`}>
@@ -26,7 +26,7 @@ const BlogCard = ({ title, date, readTimeTag, excerpt, tag, post }) => {
                   </Link> */}
           </div>
           <div className="card-info">
-            <Link href="/blog-detail">
+            <Link href={slug}>
               <h4 className="color-brand-1">{title}</h4>
             </Link>
             <div className="mb-25 mt-10">
@@ -34,7 +34,7 @@ const BlogCard = ({ title, date, readTimeTag, excerpt, tag, post }) => {
                 <Date dateString={date} />
               </span>
               <span className="font-xs color-grey-500 icon-read">
-                {/* {readTimeTag} */}2 Min Read
+                {/* {readTimeTag} */}2 min read
               </span>
             </div>
             <div
