+++ lib/posts.js
@@ -39,7 +39,67 @@ export async function getAllPosts(endCursor = null, numberOfPosts = 12) {
     }`,
   };
   const data = await graphqlRequest(query);
-  console.log(data);
   const allPosts = await data.data.posts;
   return allPosts;
+}
+
+export async function getPostData(slug) {
+  const query = {
+    query: `query singlePost {
+      post(
+        id: "${slug}",
+        idType: SLUG
+      ) {
+        date
+        title(format: RENDERED)
+        content(format: RENDERED)
+        excerpt(format: RENDERED)
+        modified
+        slug
+        featuredImage {
+          node {
+            mediaDetails {
+              sizes {
+                height
+                sourceUrl
+                width
+              }
+            }
+          }
+        }
+        categories {
+          nodes {
+            name
+            slug
+          }
+        }
+        author {
+          node {
+            name
+            slug
+          }
+        }
+      }
+    }`,
+  };
+
+  const resJson = await graphqlRequest(query);
+  const postData = resJson.data.post;
+  return postData;
+}
+
+export async function getPostSlug() {
+  const query = {
+    query: `query postSlug {
+      posts {
+        nodes {
+          slug
+        }
+      }
+    }`,
+  };
+
+  const resJson = await graphqlRequest(query);
+  const slugs = resJson.data.posts.nodes;
+  return slugs;
 }
